Update offering validator tests
