+++ src/wstore/asset_manager/test/product_validator_test_data.py
@@ -1167,278 +1167,3 @@ EMPTY_CHARS_PRODUCT = {
         }
     ],
 }
-
-BASIC_OFFERING = {
-    "id": "3",
-    "href": "http://catalog.com/offerin3",
-    "isBundle": False,
-    "name": "TestOffering",
-    "version": "1.0",
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [
-        {
-            "name": "plan",
-            "priceType": "one time",
-            "price": {
-                "taxIncludedAmount": {
-                    "unit": "EUR",
-                    "value": "1.0"
-                }
-            },
-        }
-    ],
-}
-
-OPEN_OFFERING = {
-    "id": "3",
-    "href": "http://catalog.com/offerin3",
-    "isBundle": False,
-    "name": "TestOffering",
-    "version": "1.0",
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [{
-        "id": "urn:priceid:1234",
-        "name": "open",
-        "description": "Open offering"
-    }],
-}
-
-ZERO_OFFERING = {
-    "id": "3",
-    "href": "http://catalog.com/offerin3",
-    "isBundle": False,
-    "name": "TestOffering",
-    "version": "1.0",
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [
-        {
-            "name": "plan",
-            "priceType": "one time",
-            "price": {
-                "taxIncludedAmount": {
-                    "unit": "EUR",
-                    "value": "0"
-                }
-            },
-        }
-    ],
-}
-
-FREE_OFFERING = {
-    "isBundle": False,
-    "name": "TestOffering",
-    "version": "1.0",
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-}
-
-CUSTOM_PRICING_OFFERING = {
-    "id": "3",
-    "href": "http://catalog.com/offerin3",
-    "isBundle": False,
-    "name": "TestOffering",
-    "version": "1.0",
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [
-        {
-            "name": "The custom pricing",
-            "priceType": "custom",
-            "description": "Custom pricing description"
-        }
-    ],
-}
-
-CUSTOM_PRICING_OFFERING_MULTIPLE = {
-    "id": "3",
-    "href": "http://catalog.com/offerin3",
-    "isBundle": False,
-    "name": "TestOffering",
-    "version": "1.0",
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [
-        {
-            "name": "The custom pricing",
-            "priceType": "custom",
-            "description": "Custom pricing description"
-        },
-        {
-            "name": "The custom pricing 2",
-            "priceType": "custom",
-            "description": "Custom pricing description"
-        }
-    ],
-}
-
-CUSTOM_MULTIPLE = {
-    "id": "3",
-    "href": "http://catalog.com/offerin3",
-    "isBundle": False,
-    "name": "TestOffering",
-    "version": "1.0",
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [
-        {
-            "name": "The custom pricing",
-            "priceType": "custom",
-            "description": "Custom pricing description"
-        },
-        {
-            "name": "plan",
-            "priceType": "one time",
-            "price": {
-                "taxIncludedAmount": {
-                    "unit": "EUR",
-                    "value": "1"
-                }
-            },
-        }
-    ],
-}
-
-BUNDLE_OFFERING = {
-    "isBundle": True,
-    "name": "TestOffering",
-    "version": "1.0",
-    "productSpecification": {},
-    "bundledProductOffering": [{"id": "6"}, {"id": "7"}],
-}
-
-OPEN_BUNDLE = {
-    "isBundle": True,
-    "name": "TestOffering",
-    "version": "1.0",
-    "productSpecification": {},
-    "bundledProductOffering": [{"id": "6"}, {"id": "7"}],
-    "productOfferingPrice": [{"id": "urn:priceid:1234", "name": "open", "description": "Open offering"}],
-}
-
-BUNDLE_MISSING_FIELD = {
-    "isBundle": True,
-    "name": "TestOffering",
-    "version": "1.0",
-    "productSpecification": {},
-}
-
-BUNDLE_MISSING_ELEMS = {
-    "isBundle": True,
-    "name": "TestOffering",
-    "version": "1.0",
-    "productSpecification": {},
-    "bundledProductOffering": [{"id": "6"}],
-}
-
-
-MISSING_PRICETYPE = {
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [{"name": "plan", "price": {"currencyCode": "EUR"}}],
-}
-
-INVALID_PRICETYPE = {
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [{"name": "plan", "priceType": "invalid", "price": {"currencyCode": "EUR"}}],
-}
-
-MISSING_PERIOD = {
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [{
-        "name": "plan",
-        "priceType": "recurring",
-        "price": {"currencyCode": "EUR"}
-    }],
-}
-
-INVALID_PERIOD = {
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [
-        {
-            "name": "plan",
-            "priceType": "recurring",
-            "recurringChargePeriod": "invalid",
-            "price": {"currencyCode": "EUR"},
-        }
-    ],
-}
-
-MISSING_PRICE = {
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [{"name": "plan", "priceType": "recurring", "recurringChargePeriod": "monthly"}],
-}
-
-MISSING_CURRENCY = {
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [
-        {
-            "name": "plan",
-            "priceType": "recurring",
-            "recurringChargePeriod": "monthly",
-            "price": {},
-        }
-    ],
-}
-
-INVALID_CURRENCY = {
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [
-        {
-            "name": "plan",
-            "priceType": "recurring",
-            "recurringChargePeriod": "monthly",
-            "price": {
-                "taxIncludedAmount": {
-                    "unit": "invalid"
-                }
-            },
-        }
-    ],
-}
-
-MISSING_NAME = {
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [{"priceType": "recurring", "recurringChargePeriod": "monthly", "price": {}}],
-}
-
-MULTIPLE_NAMES = {
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [
-        {
-            "name": "plan",
-            "priceType": "one time",
-            "price": {
-                "taxIncludedAmount": {
-                    "unit": "EUR",
-                    "value": "1.0"
-                }
-            },
-        },
-        {
-            "name": "Plan",
-            "priceType": "one time",
-            "price": {
-                "taxIncludedAmount": {
-                    "unit": "EUR",
-                    "value": "1.0"
-                }
-            },
-        },
-    ],
-}
-
-OPEN_MIXED = {
-    "id": "3",
-    "href": "http://catalog.com/offerin3",
-    "isBundle": False,
-    "name": "TestOffering",
-    "version": "1.0",
-    "productSpecification": {"id": "20", "href": "http://catalog.com/products/20"},
-    "productOfferingPrice": [
-        {"id": "urn:price:1234", "name": "open", "description": "Open offering"},
-        {
-            "name": "single",
-            "priceType": "one time",
-            "price": {
-                "taxIncludedAmount": {
-                    "unit": "EUR",
-                    "value": "1.0"
-                }},
-        },
-    ],
-}
