+++ src/wstore/asset_manager/test/product_validator_tests.py
@@ -28,7 +28,7 @@ from django.test.testcases import TestCase
 from mock import MagicMock, call
 from parameterized import parameterized
 
-from wstore.asset_manager import offering_validator, product_validator
+from wstore.asset_manager import product_validator
 from wstore.asset_manager.errors import ProductError
 from wstore.asset_manager.test.product_validator_test_data import *
 from wstore.store_commons.errors import ConflictError
@@ -74,7 +74,6 @@ class ValidatorTestCase(TestCase):
         product_validator.settings.SITE = "http://testlocation.org/"
 
     def tearDown(self):
-        reload(offering_validator)
         reload(product_validator)
 
     def _not_existing(self):
@@ -606,337 +605,6 @@ class ValidatorTestCase(TestCase):
         self.assertEquals(0, product_validator.Resource.objects.get.call_count)
         self.assertEquals(0, product_validator.Resource.objects.create.call_count)
 
-    def _validate_offering_calls(self, offering, asset, is_digital, is_open=False, is_custom=False):
-        # Check resource retrieving if needed
-        offering_validator.Resource.objects.filter.assert_called_once_with(
-            product_id=offering["productSpecification"]["id"]
-        )
-
-        # Check offering creation
-        offering_validator.Offering.objects.create.assert_called_once_with(
-            owner_organization=self._provider,
-            name=offering["name"],
-            description="",
-            version=offering["version"],
-            is_digital=is_digital,
-            is_open=is_open,
-            asset=asset,
-            bundled_offerings=[],
-            is_custom=is_custom
-        )
-
-    def _validate_single_offering_calls(self, offering):
-        self._validate_offering_calls(offering, self._asset_instance, True)
-
-    def _validate_physical_offering_calls(self, offering):
-        self._validate_offering_calls(offering, None, False)
-
-    def _validate_bundle_offering_calls(self, offering, is_digital, is_open=False):
-        self.assertEquals(
-            [call(off_id=off["id"]) for off in offering["bundledProductOffering"]],
-            offering_validator.Offering.objects.filter.call_args_list,
-        )
-
-        # Validate offering creation
-        offering_validator.Offering.objects.create.assert_called_once_with(
-            owner_organization=self._provider,
-            name=offering["name"],
-            description="",
-            version=offering["version"],
-            is_digital=is_digital,
-            is_open=is_open,
-            is_custom=False,
-            asset=None,
-            bundled_offerings=[off[0].pk for off in self._bundles],
-        )
-
-    def _validate_bundle_digital_offering_calls(self, offering):
-        self._validate_bundle_offering_calls(offering, True)
-
-    def _validate_bundle_physical_offering_calls(self, offering):
-        self._validate_bundle_offering_calls(offering, False)
-
-    def _validate_open_offering_calls(self, offering):
-        self._validate_offering_calls(offering, self._asset_instance, True, is_open=True)
-        self.assertTrue(self._asset_instance.is_public)
-        self._asset_instance.save.assert_called_once_with()
-
-    def _validate_open_bundle_calls(self, offering):
-        self._validate_bundle_offering_calls(offering, True, is_open=True)
-
-    def _validate_custom_pricing_calls(self, offering):
-        self._validate_offering_calls(offering, self._asset_instance, True, False, True)
-
-    def _mock_product_request(self):
-        offering_validator.requests = MagicMock()
-        product = deepcopy(BASIC_PRODUCT["product"])
-        product["id"] = "20"
-        resp = MagicMock()
-        offering_validator.requests.get.return_value = resp
-        resp.json.return_value = product
-        resp.status_code = 200
-
-    def _mock_offering_bundle(self, offering, is_digital=True):
-        offering_validator.Offering = MagicMock()
-
-        self._bundles = (
-            []
-            if "bundledProductOffering" not in offering
-            else [[MagicMock(id=off["id"], is_digital=is_digital)] for off in offering["bundledProductOffering"]]
-        )
-        offering_validator.Offering.objects.filter.side_effect = self._bundles
-
-    def _non_digital_offering(self):
-        offering_validator.Resource.objects.filter.return_value = []
-
-    def _non_digital_bundle(self):
-        self._mock_offering_bundle(BUNDLE_OFFERING, is_digital=False)
-
-    def _invalid_bundled(self):
-        offering_validator.Offering.objects.filter.side_effect = None
-        offering_validator.Offering.objects.filter.return_value = []
-
-    def _open_bundled(self):
-        for bundle_resp in self._bundles:
-            for bundle in bundle_resp:
-                bundle.is_open = True
-
-    def _open_existing(self):
-        offering_responses = [bund for bund in self._bundles]
-        offering_responses.append([MagicMock(id="6"), MagicMock(id="8")])
-        offering_validator.Offering.objects.filter.side_effect = offering_responses
-
-    def _mixed_bundled_offerings(self):
-        offering_validator.Offering.objects.filter.side_effect = [
-            [MagicMock(id="6", is_digital=True)],
-            [MagicMock(id="7", is_digital=False)],
-        ]
-
-    def _catalog_api_error(self):
-        offering_validator.requests.get().status_code = 500
-
-    def _non_open_bundled(self):
-        for bundle_resp in self._bundles:
-            for bundle in bundle_resp:
-                bundle.is_open = False
-
-    @parameterized.expand(
-        [
-            ("valid_pricing", BASIC_OFFERING, _validate_single_offering_calls, None),
-            (
-                "zero_offering",
-                ZERO_OFFERING,
-                None,
-                None,
-                "Invalid price, it must be greater than zero.",
-            ),
-            (
-                "free_offering",
-                FREE_OFFERING,
-                _validate_physical_offering_calls,
-                _non_digital_offering,
-            ),
-            (
-                "bundle_offering",
-                BUNDLE_OFFERING,
-                _validate_bundle_digital_offering_calls,
-                None,
-            ),
-            (
-                "bundle_offering_non_digital",
-                BUNDLE_OFFERING,
-                _validate_bundle_physical_offering_calls,
-                _non_digital_bundle,
-            ),
-            ("open_offering", OPEN_OFFERING, _validate_open_offering_calls, None),
-            ("open_bundle", OPEN_BUNDLE, _validate_open_bundle_calls, _open_bundled),
-            ("custom_pricing", CUSTOM_PRICING_OFFERING, _validate_custom_pricing_calls, None),
-            ("custom_pricing_multiple", CUSTOM_PRICING_OFFERING_MULTIPLE, _validate_custom_pricing_calls, None),
-            (
-                "missing_type",
-                MISSING_PRICETYPE,
-                None,
-                None,
-                "Missing required field priceType in productOfferingPrice",
-            ),
-            (
-                "invalid_type",
-                INVALID_PRICETYPE,
-                None,
-                None,
-                "Invalid priceType, it must be one time, recurring, usage, or custom",
-            ),
-            (
-                "missing_charge_period",
-                MISSING_PERIOD,
-                None,
-                None,
-                "Missing required field recurringChargePeriod for recurring priceType",
-            ),
-            (
-                "invalid_period",
-                INVALID_PERIOD,
-                None,
-                None,
-                "Unrecognized recurringChargePeriod: invalid",
-            ),
-            (
-                "missing_price",
-                MISSING_PRICE,
-                None,
-                None,
-                "Missing required field price in productOfferingPrice",
-            ),
-            (
-                "missing_currency",
-                MISSING_CURRENCY,
-                None,
-                None,
-                "Missing currency code in price",
-            ),
-            (
-                "invalid_currency",
-                INVALID_CURRENCY,
-                None,
-                None,
-                "Unrecognized currency: invalid",
-            ),
-            (
-                "missing_name",
-                MISSING_NAME,
-                None,
-                None,
-                "Missing required field name in productOfferingPrice",
-            ),
-            (
-                "multiple_names",
-                MULTIPLE_NAMES,
-                None,
-                None,
-                "Price plans names must be unique (Plan)",
-            ),
-            (
-                "bundle_missing",
-                BUNDLE_MISSING_FIELD,
-                None,
-                None,
-                "Offering bundles must contain a bundledProductOffering field",
-            ),
-            (
-                "bundle_invalid_number",
-                BUNDLE_MISSING_ELEMS,
-                None,
-                None,
-                "Offering bundles must contain at least two bundled offerings",
-            ),
-            (
-                "bundle_inv_bundled",
-                BUNDLE_OFFERING,
-                None,
-                _invalid_bundled,
-                "The bundled offering 6 is not registered",
-            ),
-            (
-                "bundle_mixed",
-                BUNDLE_OFFERING,
-                None,
-                _mixed_bundled_offerings,
-                "Mixed bundle offerings are not allowed. All bundled offerings must be digital or physical",
-            ),
-            (
-                "open_mixed",
-                OPEN_MIXED,
-                None,
-                None,
-                "Open offerings cannot include price plans",
-            ),
-            (
-                "custom_pricing_error",
-                CUSTOM_MULTIPLE,
-                None,
-                None,
-                "Custom pricing offerings cannot include processed price plans",
-            ),
-            (
-                "open_multiple_offers",
-                OPEN_OFFERING,
-                None,
-                _open_existing,
-                "Assets of open offerings cannot be monetized in other offerings",
-            ),
-            (
-                "open_non_digital",
-                OPEN_OFFERING,
-                None,
-                _non_digital_offering,
-                "Non digital products cannot be open",
-            ),
-            (
-                "open_bundle_mixed",
-                OPEN_BUNDLE,
-                None,
-                _non_open_bundled,
-                "If a bundle is open all the bundled offerings must be open",
-            ),
-        ]
-    )
-    def test_create_offering_validation(self, name, offering, checker, side_effect, msg=None):
-        self._mock_validator_imports(offering_validator)
-        offering_validator.Resource.objects.filter.return_value = [self._asset_instance]
-
-        self._mock_product_request()
-        self._mock_offering_bundle(offering)
-
-        # Add the extra call for open search
-        offering_responses = [bund for bund in self._bundles]
-        offering_responses.append([])
-        offering_validator.Offering.objects.filter.side_effect = offering_responses
-
-        if side_effect is not None:
-            side_effect(self)
-
-        error = None
-        try:
-            validator = offering_validator.OfferingValidator()
-            validator.validate("create", self._provider, offering)
-        except Exception as e:
-            error = e
-
-        if msg is not None:
-            self.assertTrue(isinstance(error, ValueError))
-            self.assertEquals(msg, str(error))
-        else:
-            self.assertEquals(error, None)
-
-            # Validate calls
-            checker(self, offering)
-
-    def test_offering_attachment(self):
-        offering = MagicMock()
-        offering_validator.Offering = MagicMock()
-        offering_validator.Offering.objects.filter.return_value = [offering]
-
-        validator = offering_validator.OfferingValidator()
-        validator.validate("attach", self._provider, BASIC_OFFERING)
-
-        self.assertEquals(BASIC_OFFERING["href"], offering.href)
-        self.assertEquals(BASIC_OFFERING["id"], offering.off_id)
-
-        offering.save.assert_called_once_with()
-
-    def test_offering_attachment_missing(self):
-        offering_validator.Offering = MagicMock()
-        offering_validator.Offering.objects.filter.return_value = []
-
-        error = None
-        try:
-            validator = offering_validator.OfferingValidator()
-            validator.validate("attach", self._provider, BASIC_OFFERING)
-        except ValueError as e:
-            error = e
-
-        self.assertEquals("The specified offering has not been registered", str(error))
-
     def _mock_non_attached(self):
         self._asset_instance.product_id = None
 
@@ -1019,54 +687,3 @@ class ValidatorTestCase(TestCase):
         validator.validate(action, self._provider, data)
 
         test_validator(self)
-
-    def _mock_offering_update(self):
-        self._mock_validator_imports(offering_validator)
-        offering_validator.Resource.objects.filter.return_value = [self._asset_instance]
-
-        self._mock_product_request()
-
-        offering_validator.Offering = MagicMock()
-        offering_validator.Offering.objects.filter.return_value = []
-
-    def test_update_offering_validator(self):
-        self._mock_offering_update()
-
-        validator = offering_validator.OfferingValidator()
-        validator.validate("update", self._provider, BASIC_OFFERING)
-
-        # Validate calls
-        self.assertEquals(0, offering_validator.Offering.objects.filter.call_count)
-        self.assertFalse(self._asset_instance.is_public)
-        self._asset_instance.save.assert_called_once_with()
-
-    def test_update_open_offering(self):
-        self._mock_offering_update()
-
-        validator = offering_validator.OfferingValidator()
-        validator.validate("update", self._provider, OPEN_OFFERING)
-
-        # Validate calls
-        offering_validator.Offering.objects.filter.assert_called_once_with(asset=self._asset_instance)
-        self.assertTrue(self._asset_instance.is_public)
-        self._asset_instance.save.assert_called_once_with()
-
-    def test_update_open_offering_multiple_error(self):
-        self._mock_offering_update()
-
-        offering_validator.Offering.objects.filter.return_value = [
-            MagicMock(),
-            MagicMock(),
-        ]
-        error = None
-        try:
-            validator = offering_validator.OfferingValidator()
-            validator.validate("update", self._provider, OPEN_OFFERING)
-        except Exception as e:
-            error = e
-
-        self.assertTrue(isinstance(error, ValueError))
-        self.assertEqual(
-            str(error),
-            "Assets of open offerings cannot be monetized in other offerings",
-        )
